optimize navbar component
